完成员工管理模块

+++ src/main/java/com/example/demo/entity/RewardsPunishment.java
@@ -6,8 +6,8 @@ import lombok.Data;
 @Data
 public class RewardsPunishment {
 
-  private long id;
-  private long employeeNumber;
+  private Integer id;
+  private Integer employeeNumber;
   private String type;
   private String reason;
   private double money;
